+++ database/migrations/2026_07_04_000000_create_image_categories_table.php
@@ -15,9 +15,4 @@ return new class () extends Migration {
             $table->timestamps();
         });
     }
-
-    public function down(): void
-    {
-        Schema::dropIfExists('image_categories');
-    }
 };