+++ src/Database/Factories/ImageCategoryFactory.php
@@ -20,7 +20,7 @@ class ImageCategoryFactory extends Factory
     public function definition(): array
     {
         return [
-            'name' => fake()->unique()->words(2, true),
+            'name' => $this->faker->unique()->words(2, true),
         ];
     }
 }